+++ lib/pghero/tasks.rb
@@ -5,4 +5,9 @@ namespace :pghero do
   task capture_query_stats: :environment do
     PgHero.capture_query_stats
   end
+
+  desc "autoindex"
+  task autoindex: :environment do
+    PgHero.autoindex_all(create: true)
+  end
 end