+++ test/explain_test.rb
@@ -2,17 +2,17 @@ require_relative "test_helper"
 
 class ExplainTest < Minitest::Test
   def setup
-    User.delete_all
+    City.delete_all
   end
 
   def test_explain
-    User.create!
-    PgHero.explain("ANALYZE DELETE FROM users")
-    assert_equal 1, User.count
+    City.create!
+    PgHero.explain("ANALYZE DELETE FROM cities")
+    assert_equal 1, City.count
   end
 
   def test_explain_multiple_statements
-    User.create!
-    assert_raises(ActiveRecord::StatementInvalid) { PgHero.explain("ANALYZE DELETE FROM users; DELETE FROM users; COMMIT") }
+    City.create!
+    assert_raises(ActiveRecord::StatementInvalid) { PgHero.explain("ANALYZE DELETE FROM cities; DELETE FROM cities; COMMIT") }
   end
 end